add cli version test
