+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 addopts = -v -ra --cov
+python_paths = src/
