--- pyvenv.cfg
@@ -1,3 +1,0 @@
-home = C:\Users\HANSBK\AppData\Local\Programs\Python\Python38
-include-system-site-packages = false
-version = 3.8.3
