+++ requirements-test.txt
@@ -1,2 +1,2 @@
---editable .[test]
 --requirement requirements.txt
+pytest-pythonpath
