--- src/example/__main__.py
@@ -1,5 +1,0 @@
-"""Code to run if this package is used as a Python module."""
-
-from .example import main
-
-main()
