--- src/example/_version.py
@@ -1,2 +1,0 @@
-"""This file defines the version of this module."""
-__version__ = "0.0.1"
