--- src/example/data/secret.txt
@@ -1,1 +1,0 @@
-Three may keep a secret, if two of them are dead.
